Update to WebFlux
Previously the implementation of this sample was based on Spring WebMVC. While
this was acceptable, the advent of Spring WebFlux means that the same
implementation can now be done in a non-blocking way for both incoming and
outgoing communication.  This makes the sample massively more efficient and
performant.  In addition, it enables multi-part requests and responses to pass
through transparently.